apps/requirements.txt
fastapi
uvicorn
websockets
python-multipart
numpy
opencv-python-headless
easyocr
torch
pillow
requests
